pyproject.toml
[project]
name = "idp-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.63.0",
    "instructor>=1.14.5",
    "jsonref>=1.1.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
]
